requirements.txt
Flask==2.0.1
pymongo[srv]==4.3.3
firebase==3.0.1
python-jwt==4.0.0
gcloud==0.18.3
sseclient==0.0.27
pycryptodome==3.18.0
requests-toolbelt==0.10.1
urllib3==1.26.15
google-auth==2.22.0
google-auth-oauthlib==1.0.0